# swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "info": {
        "title": "Concesionarios API",
        "version": "1.0.2"
    },
    "definitions": {
        "Concesionarios": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "direccion": {
                    "type": "string"
                },
                "Coche": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "modelo": {
                            "type": "string"
                        },
                        "cv": {
                            "type": "integer"
                        },
                        "precio": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "Coche": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "modelo": {
                    "type": "string"
                },
                "cv": {
                    "type": "integer"
                },
                "precio": {
                    "type": "integer"
                }
            }
        }
    },
    "paths": {
        "/Concesionarios": {
            "get": {
                "summary": "Get all Concesionarios.",
                "responses": {
                    "200": {
                        "description": "Array with Concesionarios.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Concesionarios"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a concesionario.",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "coche",
                        "schema": {
                            "$ref": "#/definitions/Concesionarios"
                        }
                    }
                ]
            }
        },
        "/Concesionarios/{id}": {
            "get": {
                "summary": "Get one concesionario by ID.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fetch one concesionario",
                        "schema": {
                            "$ref": "#/definitions/Concesionarios"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update one concesionario by ID.",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true
                    },
                    {
                        "in": "body",
                        "name": "concesionario",
                        "schema": {
                            "$ref": "#/definitions/Concesionarios"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{message: 'ok'}"
                    }
                }
            },
            "delete": {
                "summary": "Delete one concesionario by ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{message: 'ok'}"
                    }
                }
            }
        },
        "/Concesionarios/{id}/coches": {
            "get": {
                "summary": "Get all coches of one concesionario.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array with coches of one concesionario.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Coche"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a coche in a concesionario.",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true
                    },
                    {
                        "in": "body",
                        "name": "coche",
                        "schema": {
                            "$ref": "#/definitions/Coche"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{message: 'ok'}"
                    }
                }
            }
        },
        "/Concesionarios/{id}/coches/{cocheId}": {
            "get": {
                "summary": "Get one coche of one concesionario by ID.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "cocheId",
                        "type": "integer",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fetch one concesionario",
                        "schema": {
                            "$ref": "#/definitions/Coche"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update one coche of one concesionario by ID.",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "cocheId",
                        "type": "integer",
                        "required": true
                    },
                    {
                        "in": "body",
                        "name": "coche",
                        "schema": {
                            "$ref": "#/definitions/Coche"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{message: 'ok'}"
                    }
                }
            },
            "delete": {
                "summary": "Delete one coche of one concesionario by ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "cocheId",
                        "type": "integer",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{message: 'ok'}"
                    }
                }
            }
        }
    }
}